Add import and export scripts
These are a workaround for getting content into and out of the system
Kolibri instance. The normal way of exporting from the UI doesn't work
because most of the filesystem isn't available in the sandbox and the
system instance is running as a different user.

The export script copies the entire content directory and then creates a
database and channel listing. The intention is that this would be saved
to a USB drive. The import script starts a temporary Kolibri server
using the exported content and then tells the system Kolibri instance to
import from it.

https://phabricator.endlessm.com/T33448