tmpfiles: Don't recursively adjust permissions
Previously, at startup all files and directories below /var/lib/kolibri
and /var/lib/endless-key would be adjusted to be owned by the 'kolibri'
user and group, and to have mode 0755 (readable and executable by all;
writable by the owner).

Since
https://github.com/endlessm/eos-image-builder/commit/2f47bc972bddc7ded93be9a26f68c68cde427441
the change of ownership is unnecessary:

- All pre-loaded resources in new images will be owned by the correct
  user and group
- All existing systems, installed from images where the files were owned
  by root, have booted at least once and so have had the ownership
  updated
- The systemwide daemons run as the 'kolibri' user and group so any new
  files and directories created during use of Kolibri/Endless Key are
  owned by the correct user and group

Similarly we can assume that all preinstalled files are writable by
their owner, and that directories are executable as well as readable.

It is not necessary to mark files as executable: they are not executable
files, they are data. And doing so causes issues with live-booted
systems without persistant storage, where /var is an overlayfs with
upperdir and workdir stored on a tmpfs: to add the writable bit, the
files must be copied up to the writable directory. This quickly
fills the tmpfs, which can only occupy 50% of physical RAM; and the
system fails to boot.

Remove the Z line. Files on existing systems will keep the executable
bit – this is weird but harmless.

https://phabricator.endlessm.com/T35498

-----

You may ask yourself: why does overlayfs need to copy the file
contents to the upper directory just to change 3 bits in the
permissions?

From the overlayfs documentation
<https://www.kernel.org/doc/html/latest/filesystems/overlayfs.html>:

> When a file in the lower filesystem is accessed in a way the requires
> write-access, such as opening for write access, changing some metadata
> etc., the file is first copied from the lower filesystem to the upper
> filesystem (copy_up).

There is a special mode that can be used to avoid this:

> When the “metacopy” feature is enabled, overlayfs will only copy up
> metadata (as opposed to whole file), when a metadata specific
> operation like chown/chmod is performed. Full file will be copied up
> later when file is opened for WRITE operation.

This feature can be enabled in 3 ways:

1. CONFIG_OVERLAY_FS_METACOPY can be set in the kernel configuration to
   enable this feature by default. We do not set this option in our
   kernel configuration.
2. The module parameter metacopy=on can be provided at module load time.
   We do not set this option at module load time.
3. The overlayfs can be mounted with metacopy=on. We do not set this
   option at mount time.

So why is this option not enabled by default? It is a security risk in
the general case:

> Do not use metacopy=on with untrusted upper/lower directories.
> Otherwise it is possible that an attacker can create a handcrafted
> file with appropriate REDIRECT and METACOPY xattrs, and gain access to
> file on lower pointed by REDIRECT. This should not be possible on
> local system as setting “trusted.” xattrs will require CAP_SYS_ADMIN.
> But it should be possible for untrusted layers like from a pen drive.

In our case the upper & lower directories are probably “trusted enough”
in the live-media case, because there are much easier ways to do
privilege escalation if you have access to the live media than tinkering
with the overlayfs.

This may be worth exploring in future but in this case we can just avoid
copying up gigabytes of data to set an unnecessary executable bit or
three.